Attempt to cleanup where the height map live view is created and done but didn't really get anywhere
